projs/demo/requirements.txt
Django
requests
django-annoying
MYSQL-python